requirements.txt
astral>=2.2
pytz>=2021.1
pillow>=9.0.0
redis>=4.0.0
typing-extensions>=4.0.0
